requirements.txt
streamlit
google-generativeai
PyMuPDF
Pillow
python-dotenv
presidio-analyzer
presidio-anonymizer
